+++ mods/BrewLAN/modules/BrewGantries/mod_info.lua
@@ -1,4 +1,4 @@
-name = "BrewLAN: Gantry Homogeniser"
+name = "Gantry Homogeniser"
 uid = "25D57D85-7D84-27HT-A501-BR3WGANTH0000"
 version = 0.0
 copyright = "© 2018 Sean Wheeldon"